anchor_project/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-sse-client==0.2.1
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==41.0.7
Django==3.2.25
django-cors-headers==3.14.0
django-environ==0.11.2
django-model-utils==4.5.1
django-polaris==2.5.0
djangorestframework==3.15.1
frozenlist==1.5.0
idna==3.10
mnemonic==0.20
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pytz==2024.2
requests==2.32.3
requests-sse==0.3.2
sqlparse==0.4.4
stellar-sdk==10.0.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==5.3.0
xdrlib3==0.1.1
yarl==1.18.3
